Move sparse matrix serialization to separate files
